requirements.txt
typer
azure-identity
azure-mgmt-network
ipaddress
rich 
streamlit 
azure-mgmt-resource 